+++ spline/interp2Dcomp.h
@@ -16,7 +16,7 @@ int interp2Dcomp_eval_B(real* B, interp2D_data* str, real r, real z);
 #pragma omp declare simd linear(i) uniform(B, str)
 int interp2Dcomp_eval_B_SIMD(int i, real B[NSIMD], interp2D_data* str, real r, real z);
 #pragma omp declare simd uniform(str)
-int interp2Dcomp_eval_dB(real* B_dB, interp2D_data* str, real r, real z);
+long int interp2Dcomp_eval_dB(real* B_dB, interp2D_data* str, real r, real z);
 #pragma omp declare simd linear(i) uniform(B_dB, str)
 int interp2Dcomp_eval_dB_SIMD(int i, real B_dB[6][NSIMD], interp2D_data* str, real r, real z);
 void interp2Dcomp_free(interp2D_data* str);
